Minor rearrangement
